chore: 여러 자잘한 수정사항 refactor: userdb.js -> service.js로 변경 - 공통 서비스 로직 관리 	renamed:    scripts/userdb.js -> scripts/service.js 	modified:   signin/app.js 	modified:   signup/app.js
